[build] switch to xbstrap for dependency building and a script image creation
From now on, xbstrap will handle building and managing third-party code
that needs to be built. This simplifies things a lot, especially in
terms of maintenance.

The new script for building images is located at util/builder. It takes
a configuration file written in JSON and prints a bash script producing
the specified image to stdout.